test: Update server configuration for deployment on Vercel
