+++ package.json
@@ -10,12 +10,15 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@vercel/node": "^3.2.23",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "minimist": "^1.2.8",
     "node-fetch": "^2.6.7",
-    "redis": "^4.5.1"
+    "redis": "^4.5.1",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.6.3"
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
