+++ src/types/express/index.d.ts
@@ -1,0 +1,10 @@
+import { RedisWrapper } from "../../redisClient";
+
+declare global {
+  namespace Express {
+    interface Request {
+      redisClient: RedisWrapper | null;
+      disableCache: boolean;
+    }
+  }
+}
