+++ tsconfig.json
@@ -11,7 +11,9 @@
       "*": ["node_modules/*"]
     },
     "typeRoots": ["./node_modules/@types", "./types"],
-    "outDir": "./dist"
+    "rootDir": "./src",
+    "outDir": "./dist",
   },
+  "include": ["src/**/*.ts"],
   "exclude": ["node_modules"]
 }
