--- types/express/index.d.ts
@@ -1,10 +1,0 @@
-import { RedisWrapper } from "../../redisClient";
-
-declare global {
-    namespace Express {
-        interface Request {
-            redisClient: RedisWrapper | null;
-            disableCache: boolean;
-        }
-    }
-}
